Few changes

+++ ProductService.BLL/DTO/ProductResponse.cs
@@ -1,6 +1,6 @@
 ﻿namespace ProductService.BLL.DTO;
 
-public record ProductResponse(Guid ProductID, string ProductName, CategoryOptions Category, double? UnitPrice, int? QuantityInStock)
+public record ProductResponse(Guid ProductID, string ProductName, CategoryOptions Category, decimal? UnitPrice, int? QuantityInStock)
 {
   public ProductResponse() : this(default, default, default, default, default)
   {
